Issue #20; Added TravellingFragment
I created TravellingFragment.java with the appropriate layout. I defined onClick events for navigation through the travelling section of the application. I also added the option to the navigation drawer menu which redirects the user to the newly created fragment. TODO: show real-time location in TravellingFragment and try to show the distance to our destination.