Initializes #70; Weather data download using Retrofit
As data needs to be structured and easily available when dealing with charts, I decided to switch to Retrofit for getting OpenWeatherMap data. That includes implementing interface for Retrofit and all necessary classes that reflect OpenWeatherMap One Call API response. The data should now be easily available to anyone who makes a call towards the API using Retrofit and 'OpenWeatherMapAPI' interface.