Fixed a small bug when continuing a trip.
The problem was that the destination sometimes wouldn't show up on the map after restarting the application and going to On the go section immediately. The bug was related to lack of synchronization and was easy to fix - I just made the app wait for the map to be ready and to show the tdestination then.