Fixed bug similar to the ones before
Again, a bug caused by faulty execution sequence. I think that we will need to review every class and make ourselves sure that there are not any unpredictable problems that can be caused by different time lengths of tasks that rely on resources that often vary in their response times.