Removed imports causing errors
I removed imports that were causing errors in 'CityCoordConversionTest.java'. As the test was running fine anyways because the imports weren't used, the mistake went by unnoticed. I didn't think that creating a new branch just to remove a couple of unused imports would be appropriate, so I am commiting directly to main.